Introduce untrusted transition support for newfs_msdos execution
Initial VOLD implementation did not execute a SEPOLICY transition
for the newfs_msdos binary, and instead relied on the default
"system_file" label to stay in the "vold" domain.

Update VOLD implementation to undergo a transition on newfs_msdos
execution based on the "fsck_untrusted_context" option specified on
vold startup.

Change-Id: I439c5956774789b9b5c770cde1e0341be17077ff
CRs-Fixed: 2789815